Changes to time granularity. Works fine for cars, there is an issue with the traffic lights. Congestion rate creates more cars but, at the moment, doesn't remove any. Emergency cars not yet produced. Changed the delay of traffic lights from float to int.

+++ Traffic_Control/src/Objects/TrafficLights.java
@@ -21,7 +21,7 @@ public class TrafficLights extends JPanel implements Runnable{
     private TrafficLights resumeNextLight;
     private boolean suspendRequest;
     private int checkFirst;
-    private long delay;
+    private int delay;
     private int signal;
     private int numberOfWays;
 
@@ -33,7 +33,7 @@ public class TrafficLights extends JPanel implements Runnable{
      * @param y_coordinate
      * @param rotation
      */
-    public TrafficLights(int x_coordinate, int y_coordinate, int numberOfWays, int signal, int rotation, long delay){
+    public TrafficLights(int x_coordinate, int y_coordinate, int numberOfWays, int signal, int rotation, int delay){
         this.pos_x = x_coordinate;
         this.pos_y = y_coordinate;
         this.rotates = rotation;
@@ -168,7 +168,7 @@ public class TrafficLights extends JPanel implements Runnable{
             }
     }
 
-    private long getDelay() {
+    private int getDelay() {
         return delay;
     }
 }
